test(refactor): assertion helper 분리
